+++ src/Componentes/Home/Home.tsx
@@ -3,16 +3,65 @@ import CardsContainer from 'Componentes/CardsContainer';
 import Filtrado from 'Componentes/Filtrado';
 import './Home.css';
 import { traerPets } from 'helpers';
-import Login from 'Componentes/Login/Login';
-import Logout from 'Componentes/Login/Logout';
 import { useAuth0 } from '@auth0/auth0-react';
+import Swal from 'sweetalert2';
+import { useNavigate } from 'react-router-dom';
+import axios from 'axios';
+import { type Pet } from 'redux/slices/mascotas';
 
 const Home: React.FC = () => {
-  const { isAuthenticated } = useAuth0();
-  console.log(isAuthenticated);
+  const { isAuthenticated, user, logout } = useAuth0();
+  const navigate = useNavigate();
+  const { loginWithRedirect } = useAuth0();
   useEffect((): void => {
     traerPets();
   });
+  useEffect((): void => {
+    if (isAuthenticated && user != null) {
+      const { name, email, image } = user;
+      auth0Logica(name, image, email);
+    }
+  });
+  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
+  const login = () => {
+    Swal.fire({
+      title: 'Tienes cuenta en PetsHouse?',
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: 'Login',
+      denyButtonText: `Ingresa con gmail`
+    }).then((result) => {
+      if (result.isConfirmed) {
+        navigate('/loginAPI');
+      } else if (result.isDenied) {
+        loginWithRedirect();
+      }
+    });
+  };
+  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
+  async function auth0Logica(
+    name: string | undefined,
+    image: any,
+    email: string | undefined
+  ) {
+    try {
+      await axios
+        .post<Pet[]>(`/users/userAuth0`, { name, image, email })
+        .then((res: { data: any }) => {
+          localStorage.setItem('id', res.data.id);
+          localStorage.setItem('name', res.data.name);
+          localStorage.setItem('image', res.data.image);
+          localStorage.setItem('rol', res.data.rol);
+        });
+    } catch (error) {
+      console.log(error);
+    }
+  }
+  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
+  const logoutApp = () => {
+    localStorage.clear();
+    logout({ logoutParams: { returnTo: window.location.origin } });
+  };
 
   return (
     <div>
@@ -27,8 +76,11 @@ const Home: React.FC = () => {
           <CardsContainer />
         </div>
       </div>
-      <Login />
-      <Logout />
+      {isAuthenticated ? (
+        <button onClick={logoutApp}>Logout</button>
+      ) : (
+        <button onClick={login}>Login</button>
+      )}
     </div>
   );
 };
