+++ src/Componentes/Login/Login.tsx
@@ -1,19 +1,89 @@
-import React from 'react';
-import { useAuth0 } from '@auth0/auth0-react';
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { type Pet } from 'redux/slices/mascotas';
+import Swal from 'sweetalert2';
+import axios from 'axios';
 
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const LoginButton = () => {
-  const { loginWithRedirect } = useAuth0();
+const Login: React.FC = () => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const navigate = useNavigate();
 
+  async function handleSubmit(
+    e: React.FormEvent<HTMLFormElement>
+  ): Promise<void> {
+    e.preventDefault();
+    try {
+      await axios
+        .get<Pet[]>(`/users/login/${email}/${password}`)
+        .then((res: { data: any }) => {
+          if (typeof res.data === 'string') {
+            Swal.fire({
+              title: '¡Error!',
+              text: res.data,
+              icon: 'error',
+              confirmButtonText: 'Intentar de nuevo'
+            });
+          } else {
+            const { id, name, image, rol } = res.data;
+            localStorage.setItem('id', id);
+            localStorage.setItem('name', name);
+            localStorage.setItem('image', image);
+            localStorage.setItem('rol', rol);
+            navigate('/');
+          }
+        });
+    } catch (error) {
+      console.log(error);
+    }
+  }
   return (
-    <button
-      onClick={() => {
-        loginWithRedirect();
-      }}
-    >
-      Log In
-    </button>
+    <>
+      <form
+        onSubmit={(e) => {
+          handleSubmit(e);
+        }}
+      >
+        <div>
+          <label htmlFor="email">Correo electronico</label>
+          <input
+            type="text"
+            name="email"
+            placeholder="yourmail@email.com"
+            autoComplete="off"
+            onChange={({ target }) => {
+              setEmail(target.value);
+            }}
+            required
+            value={email}
+          />
+          <br />
+          <label htmlFor="password">Contraseña</label>
+          <input
+            type="password"
+            name="password"
+            onChange={({ target }) => {
+              setPassword(target.value);
+            }}
+            required
+            value={password}
+          />
+        </div>
+        <div>
+          <button type="submit">Ingresar</button>
+        </div>
+      </form>
+      <button>olvide mi contraseña</button>
+      <button>crear cuenta</button>
+      <button
+        onClick={() => {
+          navigate('/');
+        }}
+      >
+        Home
+      </button>
+    </>
   );
 };
 
-export default LoginButton;
+export default Login;
