+++ src/Componentes/Login/LoginAuth0.tsx
@@ -2,19 +2,18 @@ import React from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const LogoutButton = () => {
-  const { logout } = useAuth0();
+const LoginButton = () => {
+  const { loginWithRedirect } = useAuth0();
 
   return (
     <button
       onClick={() => {
-        logout({ logoutParams: { returnTo: window.location.origin } });
-        localStorage.clear();
+        loginWithRedirect();
       }}
     >
-      Log Out
+      Log In
     </button>
   );
 };
 
-export default LogoutButton;
+export default LoginButton;
