+++ src/Componentes/Publicaciones/index.tsx
@@ -2,8 +2,10 @@ import Card from 'Componentes/Card';
 import { getUserPublications } from 'helpers';
 import { useCustomSelector } from 'hooks/redux';
 import React, { useEffect } from 'react';
+import { useAuth0 } from '@auth0/auth0-react';
 
 const Publicaciones: React.FC = () => {
+  const { isAuthenticated } = useAuth0();
   // const idUser = Number(localStorage.getItem('id'));
   const idUser = 1;
   useEffect((): void => {
@@ -12,7 +14,7 @@ const Publicaciones: React.FC = () => {
 
   const publications = useCustomSelector((state) => state.pets.publications);
 
-  if (idUser > 0) {
+  if (isAuthenticated) {
     return (
       <div>
         {publications.map((pet, index) => {
